# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MatriConnect Matrimony API",
        "description": "Full-stack matrimony platform API. All protected endpoints require Bearer token from Sanctum.",
        "contact": {
            "email": "admin@matriconnect.com"
        },
        "license": {
            "name": "MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Primary API Server"
        }
    ],
    "paths": {
...
        "/api/v1/admin/users": {
            "get": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "List all users with pagination",
                "operationId": "b0b00cf5db5258ff35b911db295a30da",
                "parameters": [
                    {
                        "name": "search",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Users list"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}/ban": {
            "put": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Ban or unban a user",
                "operationId": "83fed388a4d5d92865f0e0b255709240",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "is_banned": {
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "reason": {
                                        "type": "string",
                                        "example": "Violation of terms"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User ban status updated"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/v1/admin/users/{id}/verify": {
            "put": {
                "tags": [
                    "Admin - Users"
                ],
                "summary": "Verify a user profile",
                "operationId": "d84cf165f181bb7bf226c472cd23bcc4",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User verified"
                    },
                    "404": {
                        "description": "User not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/v1/auth/password/forgot": {
            "post": {
                "tags": [
                    "Password Reset"
                ],
                "summary": "Send password reset link to email",
                "operationId": "db8d6cb1928edd18c9a768f0d23125b7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Reset link sent (or silently ignored if email not found)"
                    },
                    "422": {
                        "description": "Validation error"
                    },
                    "429": {
                        "description": "Too many requests"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Laravel Sanctum token. Obtain via POST /api/v1/auth/login",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}